Renamed quantum_info module to qip
