+++ qutip/qip/__init__.py
@@ -1,0 +1,5 @@
+from qutip.qip.register import Register
+from qutip.qip.circuit import Circuit, QubitCircuit, Gate
+from qutip.qip.states import *
+from qutip.qip.gates import *
+from qutip.qip.qft import qft, qft_steps
