+++ qutip/qip/qcircuit_latex.py
