+++ qutip/qip/register.py
@@ -35,8 +35,8 @@ from qutip.qobj import *
 from qutip.states import *
 from qutip.operators import *
 from qutip.tensor import tensor
-from qutip.quantum_info.gates import *
-from qutip.quantum_info.utils import _reg_str2array
+from qutip.qip.gates import *
+from qutip.qip.utils import _reg_str2array
 
 
 class Register(Qobj):
