+++ qutip/qip/utils.py
