--- qutip/quantum_info/__init__.py
@@ -1,5 +1,0 @@
-from qutip.quantum_info.register import Register
-from qutip.quantum_info.circuit import Circuit, QubitCircuit, Gate
-from qutip.quantum_info.states import *
-from qutip.quantum_info.gates import *
-from qutip.quantum_info.qft import qft, qft_steps
