+++ qutip/tests/test_qpt.py
@@ -36,7 +36,7 @@ from numpy.testing import assert_, run_module_suite
 from numpy.linalg import norm
 
 from qutip import *
-from qutip.quantum_info.gates import *
+from qutip.qip.gates import *
 
 
 def test_qpt_snot():
